Initial Files
The files included are the database and model of the application and a module file called SQLTools.py, the PPModel.py file contains the main function of the application while the SQLTools.py contains relevant modules that is used by PPModel.py to function. PPDatabse contains the SQL database of the file and is used to store the data sent by the PPModel.py. Currently, the only functions are registering for the users and employees.